requirements.txt
django
django-bootstrap4
django-cropperjs
django-login-otp
pint
loguru
markdown
humanize